Created separate Python files for auto-creation
Added the following python files:
 - home_page.py
 - champion_page.py (need to fix the logic to select champions)
 - year_page.py
 - team_page.py
 - page_header.py

Modified the following python files:
 - constants.py
 - document.py

In all of the above, as well as auto-creation, modified the file paths to exclude 'fantasy-football-website', since now the all python files should be run from the fantasy-football-website directory
