+++ auto-creation.py
@@ -2,23 +2,24 @@ import dominate
 from dominate.tags import *
 import pandas as pd
 import numpy as np
-import sqlite3
 from matplotlib import pyplot as plt
 
-from python import functions, constants
-
-ROOT = '/fantasy-football-website/'
-
-MATCHUP_DATA = pd.read_csv('fantasy-football-website/database/fantasy-football-matchup-data.csv')
-GAME_DATA = pd.read_csv('fantasy-football-website/database/fantasy-football-game-data.csv')
-TEAMS = GAME_DATA.loc[~ GAME_DATA['Playoff Flag'], 'Team'].unique()
-YEARS = GAME_DATA['Year'].unique()
-YEARS_WEEKS = [(year, GAME_DATA.loc[(GAME_DATA['Year'] == year) & (GAME_DATA['Playoff Flag'] == False), 'Week'].max()) for year in YEARS]
+from python import functions, constants, home_page, team_page, year_page, champion_page
 
 
 
+# ROOT = '/fantasy-football-website/'
 
-def header(active_year: int = None) -> div:
+# MATCHUP_DATA = pd.read_csv('fantasy-football-website/database/fantasy-football-matchup-data.csv')
+# GAME_DATA = pd.read_csv('fantasy-football-website/database/fantasy-football-game-data.csv')
+# TEAMS = constants.GAME_DATA.loc[constants.GAME_DATA['Playoff Flag'] == 0, 'Team'].unique()
+# YEARS = constants.GAME_DATA['Year'].unique()
+# YEARS_WEEKS = [(year, constants.GAME_DATA.loc[(constants.GAME_DATA['Year'] == year) & (constants.GAME_DATA['Playoff Flag'] == False), 'Week'].max()) for year in YEARS]
+
+
+
+
+def page_header(active_year: int = None) -> div:
     '''
     Creates the blue header div, adds in the logo and heading at the top
 
@@ -35,7 +36,7 @@ def header(active_year: int = None) -> div:
         div container for the entire header
     '''
     container = div(_class='header')
-    logo = img(src=f'{ROOT}Assets/Fantasy-Football-App-LOGO.png', height=85, style='float: left;padding: 0px 20px')
+    logo = img(src=f'{constants.ROOT}Assets/Fantasy-Football-App-LOGO.png', height=85, style='float: left;padding: 0px 20px')
     heading = h1('Fantasy Football Luck Scores')
 
     navbar = topnav(active_year=active_year)
@@ -65,22 +66,22 @@ def topnav(active_year: int = None) -> div:
     container = div(_class='topnav')
 
     if active_year == 'home':
-        home = div(a('Home', href=f'{ROOT}'), _class='dropdown active')
+        home = div(a('Home', href=f'{constants.ROOT}'), _class='dropdown active')
     else:
-        home = div(a('Home', href=f'{ROOT}'), _class='dropdown')
+        home = div(a('Home', href=f'{constants.ROOT}'), _class='dropdown')
     container.add(home)
 
-    for year, week in YEARS_WEEKS:
+    for year, week in constants.YEARS_WEEKS:
         if active_year == year:
             dropdown = div(_class='dropdown active')
         else:
             dropdown = div(_class='dropdown')
-        dropdown_button = a(year, href=f'{ROOT}seasons/{year}/', _class='dropdown-button')
+        dropdown_button = a(year, href=f'{constants.ROOT}seasons/{year}/', _class='dropdown-button')
         dropdown.add(dropdown_button)
 
         dropdown_content = div(_class='dropdown-content')
         for i in range(week):
-            _a = a(f'Week {i+1}', href=f'{ROOT}seasons/{year}/week-{i+1}.html')
+            _a = a(f'Week {i+1}', href=f'{constants.ROOT}seasons/{year}/week-{i+1}.html')
 
             dropdown_content.add(_a)
 
@@ -91,12 +92,12 @@ def topnav(active_year: int = None) -> div:
         team_dropdown = div(_class='dropdown active')
     else:
         team_dropdown = div(_class='dropdown')
-    team_dropdown_button = a('Teams', href=f'{ROOT}teams/', _class='dropdown-button')
+    team_dropdown_button = a('Teams', href=f'{constants.ROOT}teams/', _class='dropdown-button')
     team_dropdown.add(team_dropdown_button)
 
     team_dropdown_content = div(_class='dropdown-content')
-    for team in TEAMS:
-        _a = a(f'{team}', href=f'{ROOT}teams/{team}.html')
+    for team in constants.TEAMS:
+        _a = a(f'{team}', href=f'{constants.ROOT}teams/{team}.html')
 
         team_dropdown_content.add(_a)
 
@@ -104,9 +105,9 @@ def topnav(active_year: int = None) -> div:
     container.add(team_dropdown)
 
     if active_year == 'champion':
-        champions = div(a('Champions', href=f'{ROOT}champion.html'), _class='dropdown active')
+        champions = div(a('Champions', href=f'{constants.ROOT}champion.html'), _class='dropdown active')
     else:
-        champions = div(a('Champions', href=f'{ROOT}champion.html'), _class='dropdown')
+        champions = div(a('Champions', href=f'{constants.ROOT}champion.html'), _class='dropdown')
     container.add(champions)
 
     return container
@@ -141,7 +142,7 @@ def week_content(year: int, week: int) -> div:
     title = h1(f'{year} Week {week}')
 
     # Create scoreboard table of the matchups that week
-    scoreboard = MATCHUP_DATA.loc[(MATCHUP_DATA['Year'] == year) & (MATCHUP_DATA['Week'] == week), ['Home Team','Home Score','Away Score','Away Team']].copy()    
+    scoreboard = constants.MATCHUP_DATA.loc[(constants.MATCHUP_DATA['Year'] == year) & (constants.MATCHUP_DATA['Week'] == week), ['Home Team','Home Score','Away Score','Away Team']].copy()    
     scoreboard_div = div(_class='content-container')
     scoreboard_title = h2('Weekly Scoreboard')
     scoreboard_table = functions.df_to_table(data=scoreboard)
@@ -150,7 +151,7 @@ def week_content(year: int, week: int) -> div:
 
     # Create the weekly recap stats section
     # This temp df needs to look at only the current week
-    temp = GAME_DATA.loc[(GAME_DATA['Year'] == year) & (GAME_DATA['Week'] == week)].copy()
+    temp = constants.GAME_DATA.loc[(constants.GAME_DATA['Year'] == year) & (constants.GAME_DATA['Week'] == week)].copy()
 
 
     highest_scorer = temp.sort_values('Score', ascending=False, ignore_index=True).loc[0,['Team','Score']].values
@@ -168,7 +169,7 @@ def week_content(year: int, week: int) -> div:
     stats_div.add([stats_title, stats_list])
 
     # Create the live standings table
-    standings = functions.summary_table(GAME_DATA, year = year, week = week)
+    standings = functions.summary_table(constants.GAME_DATA, year = year, week = week)
     standings = standings.drop('Year', axis = 1)
 
     standings_div = div(_class='content-container')
@@ -188,183 +189,183 @@ def week_content(year: int, week: int) -> div:
 
 # Function which creates home page content
 # Edit this for the content on the home page
-def home_content() -> div:
-    '''
-    Function which creates the home page content
+# def home_content() -> div:
+#     '''
+#     Function which creates the home page content
 
-    Includes:
-    ---------
-    * Documentation PDF
-    '''
-    container = div(_class='content')
-    container.add(h1('Documentation'))
+#     Includes:
+#     ---------
+#     * Documentation PDF
+#     '''
+#     container = div(_class='content')
+#     container.add(h1('Documentation'))
 
-    documentation = object_('Documentation', data='/fantasy-football-website/Assets/Fantasy-Football.pdf', height='1000px', width='90%')
-    container.add(documentation)
+#     documentation = object_('Documentation', data='/fantasy-football-website/Assets/Fantasy-Football.pdf', height='1000px', width='90%')
+#     container.add(documentation)
 
-    return container
+#     return container
 
 
 # Function which creates champion page content
 # Edit this for the content on the champion page
-def champion_content() -> div:
-    '''
-    Function which creates the champion page content
+# def champion_content() -> div:
+#     '''
+#     Function which creates the champion page content
 
-    Includes:
-    ---------
-    * Summary Table for each season
-    * List of all Champions
-    '''
-    container = div(_class='content')
-    container.add(h1('League Champions'))
+#     Includes:
+#     ---------
+#     * Summary Table for each season
+#     * List of all Champions
+#     '''
+#     container = div(_class='content')
+#     container.add(h1('League Champions'))
 
-    champions = GAME_DATA.loc[(GAME_DATA['Week'] == 17) & (GAME_DATA['Win'] == 1), ['Year','Team']].values
-    data = []
-    for year, team in champions:
-        temp = GAME_DATA.loc[(GAME_DATA['Year'] == year) & (GAME_DATA['Team'] == team) & (GAME_DATA['Playoff Flag'] == False)]
-        wins = temp['Win'].sum()
-        losses = len(temp) - wins
-        record = f'{wins}-{losses}'
+#     champions = constants.GAME_DATA.loc[(constants.GAME_DATA['Week'] == 17) & (constants.GAME_DATA['Win'] == 1), ['Year','Team']].values
+#     data = []
+#     for year, team in champions:
+#         temp = constants.GAME_DATA.loc[(constants.GAME_DATA['Year'] == year) & (constants.GAME_DATA['Team'] == team) & (constants.GAME_DATA['Playoff Flag'] == False)]
+#         wins = temp['Win'].sum()
+#         losses = len(temp) - wins
+#         record = f'{wins}-{losses}'
 
-        pf = round(temp['Score'].sum(), 2)
-        avg_pf = round(pf / len(temp), 2)
-        pa = round(temp['Opp Score'].sum(), 2)
-        avg_margin = round((pf - pa) / len(temp), 2)
+#         pf = round(temp['Score'].sum(), 2)
+#         avg_pf = round(pf / len(temp), 2)
+#         pa = round(temp['Opp Score'].sum(), 2)
+#         avg_margin = round((pf - pa) / len(temp), 2)
 
-        data.append([year, team, record, pf, avg_pf, avg_margin])
+#         data.append([year, team, record, pf, avg_pf, avg_margin])
 
-    t = pd.DataFrame(data, columns=['Year','Team','Record','Points For','Avg Points For','Avg Margin'])
+#     t = pd.DataFrame(data, columns=['Year','Team','Record','Points For','Avg Points For','Avg Margin'])
 
-    summary_div = div(_class='content-container')
-    summary_title = h2('Champions Summary')
-    summary_table = functions.df_to_table(data=t)
-    summary_table['id'] = 'champion-summary-table'
-    summary_div.add([summary_title, summary_table])
+#     summary_div = div(_class='content-container')
+#     summary_title = h2('Champions Summary')
+#     summary_table = functions.df_to_table(data=t)
+#     summary_table['id'] = 'champion-summary-table'
+#     summary_div.add([summary_title, summary_table])
 
-    years_div = div(_class='champion-years')
-    for year, team in champions:
-        years_div.add(h3(f'{year} Season'))
-        years_div.add(p(team))
+#     years_div = div(_class='champion-years')
+#     for year, team in champions:
+#         years_div.add(h3(f'{year} Season'))
+#         years_div.add(p(team))
 
-    container.add(summary_div)
-    container.add(years_div)
+#     container.add(summary_div)
+#     container.add(years_div)
 
-    return container
+#     return container
 
 
-def team_content(team: str, conn: sqlite3.Connection) -> div:
-    container = div(_class='content')
-    container.add(h1(f'{team} Data'))
+# def team_content(team: str) -> div:
+#     container = div(_class='content')
+#     container.add(h1(f'{team} Data'))
 
-    seasons = [functions.summary_table(GAME_DATA, year=year) for year in YEARS]
-    seasons_df = pd.concat(seasons)
-    seasons_df = seasons_df.loc[seasons_df['Team'] == team].drop('Team', axis = 1)
+#     seasons = [functions.summary_table(constants.GAME_DATA, year=year) for year in constants.YEARS]
+#     seasons_df = pd.concat(seasons)
+#     seasons_df = seasons_df.loc[seasons_df['Team'] == team].drop('Team', axis = 1)
 
-    summary_div = div(_class='content-container')
-    summary_title = h2('Team Summary')
-    summary_table = functions.df_to_table(data=seasons_df)
-    summary_table['id'] = 'team-summary-table'
-    summary_div.add([summary_title, summary_table])
+#     summary_div = div(_class='content-container')
+#     summary_title = h2('Team Summary')
+#     summary_table = functions.df_to_table(data=seasons_df)
+#     summary_table['id'] = 'team-summary-table'
+#     summary_div.add([summary_title, summary_table])
 
-    plt.figure(figsize = (5,3), dpi = 300)
-    plt.grid(linewidth = 0.5, zorder = 0)
-    plt.plot(seasons_df['Year'], seasons_df['Luck Score'], c = constants.COLOR_DICT[team.lower()], marker = 'o', zorder = 3)
-    plt.xlabel('Season')
-    plt.ylabel('Luck Score')
-    plt.savefig(f'fantasy-football-website/Assets/Luck-Score-Year-{team}.svg', format = 'svg', bbox_inches = 'tight')
+#     plt.figure(figsize = (5,3), dpi = 300)
+#     plt.grid(linewidth = 0.5, zorder = 0)
+#     plt.plot(seasons_df['Year'], seasons_df['Luck Score'], c = constants.COLOR_DICT[team.lower()], marker = 'o', zorder = 3)
+#     plt.xlabel('Season')
+#     plt.ylabel('Luck Score')
+#     plt.savefig(f'Assets/Luck-Score-Year-{team}.svg', format = 'svg', bbox_inches = 'tight')
 
-    line_chart_div = div(_class = 'content-container')
-    line_chart_title = h2('Luck Score by Season')
-    line_chart_img = img(src = f'{ROOT}Assets/Luck-Score-Year-{team}.svg')
-    line_chart_div.add([line_chart_title, line_chart_img])
+#     line_chart_div = div(_class = 'content-container')
+#     line_chart_title = h2('Luck Score by Season')
+#     line_chart_img = img(src = f'{constants.ROOT}Assets/Luck-Score-Year-{team}.svg')
+#     line_chart_div.add([line_chart_title, line_chart_img])
     
-    container.add(summary_div)
-    container.add(line_chart_div)
+#     container.add(summary_div)
+#     container.add(line_chart_div)
 
-    return container
+#     return container
 
 
-def year_content(year: int) -> div:
-    container = div(_class='content')
-    container.add(h1(f'{year} Data'))
+# def year_content(year: int) -> div:
+#     container = div(_class='content')
+#     container.add(h1(f'{year} Data'))
 
-    data = functions.summary_table(GAME_DATA, year = year)
-    data = data.drop('Year', axis = 1)
+#     data = functions.summary_table(constants.GAME_DATA, year = year)
+#     data = data.drop('Year', axis = 1)
 
-    s = [(team, pf, pa, constants.COLOR_DICT[team.lower()]) for team, pf, pa in zip(data['Team'], data['Points For'], data['Points Against'])]
+#     s = [(team, pf, pa, constants.COLOR_DICT[team.lower()]) for team, pf, pa in zip(data['Team'], data['Points For'], data['Points Against'])]
 
-    plt.figure(figsize = (5,3), dpi = 300)
-    plt.grid(linewidth = 0.5, zorder = 0)
-    for team, pf, pa, color in s:
-        plt.scatter(pf, pa, c = color, label = team, zorder = 3, edgecolor = 'k')
-    plt.xlim(int(data['Points For'].min() - 4) - (int(data['Points For'].min()-4) % 25),
-            int(data['Points For'].max() + 4) + 25 - (int(data['Points For'].max() + 4) % 25)) # Automatically finds nearest 25 below/above for range
-    plt.ylim(int(data['Points Against'].min() - 4) - (int(data['Points Against'].min() - 4) % 25),
-            int(data['Points Against'].max() + 4) + 25 - (int(data['Points Against'].max() + 4) % 25))
-    plt.xlabel('Points For')
-    plt.ylabel('Points Against')
-    plt.legend(bbox_to_anchor=(1.05, 1), loc='upper left', fontsize = 9)
-    plt.savefig(f'fantasy-football-website/Assets/PF-vs-PA-{year}.svg', format = 'svg',bbox_inches = 'tight')
+#     plt.figure(figsize = (5,3), dpi = 300)
+#     plt.grid(linewidth = 0.5, zorder = 0)
+#     for team, pf, pa, color in s:
+#         plt.scatter(pf, pa, c = color, label = team, zorder = 3, edgecolor = 'k')
+#     plt.xlim(int(data['Points For'].min() - 4) - (int(data['Points For'].min()-4) % 25),
+#             int(data['Points For'].max() + 4) + 25 - (int(data['Points For'].max() + 4) % 25)) # Automatically finds nearest 25 below/above for range
+#     plt.ylim(int(data['Points Against'].min() - 4) - (int(data['Points Against'].min() - 4) % 25),
+#             int(data['Points Against'].max() + 4) + 25 - (int(data['Points Against'].max() + 4) % 25))
+#     plt.xlabel('Points For')
+#     plt.ylabel('Points Against')
+#     plt.legend(bbox_to_anchor=(1.05, 1), loc='upper left', fontsize = 9)
+#     plt.savefig(f'Assets/PF-vs-PA-{year}.svg', format = 'svg',bbox_inches = 'tight')
 
-    scatter_div = div(_class='content-container')
-    scatter_title = h2('PF vs PA')
-    scatter_img = img(src = f'{ROOT}Assets/PF-vs-PA-{year}.svg')
-    scatter_div.add([scatter_title, scatter_img])
+#     scatter_div = div(_class='content-container')
+#     scatter_title = h2('PF vs PA')
+#     scatter_img = img(src = f'{constants.ROOT}Assets/PF-vs-PA-{year}.svg')
+#     scatter_div.add([scatter_title, scatter_img])
 
-    summary_div = div(_class='content-container')
-    summary_title = h2('Regular Season Summary')
-    summary_table = functions.df_to_table(data=data)
-    summary_table['id'] = 'season-summary-table'
-    summary_div.add([summary_title, summary_table])
+#     summary_div = div(_class='content-container')
+#     summary_title = h2('Regular Season Summary')
+#     summary_table = functions.df_to_table(data=data)
+#     summary_table['id'] = 'season-summary-table'
+#     summary_div.add([summary_title, summary_table])
 
-    playoff_matchups = MATCHUP_DATA.loc[(MATCHUP_DATA['Year'] == year) & (MATCHUP_DATA['Playoff Flag'])].copy()
-    playoff_matchups['Playoff Round'] = (playoff_matchups['Week'] % playoff_matchups['Week'].min()) + 1
+#     playoff_matchups = constants.MATCHUP_DATA.loc[(constants.MATCHUP_DATA['Year'] == year) & (constants.MATCHUP_DATA['Playoff Flag'])].copy()
+#     playoff_matchups['Playoff Round'] = (playoff_matchups['Week'] % playoff_matchups['Week'].min()) + 1
 
-    bracket_div = div(_class='content-container')
-    bracket_title = h2('Playoff Bracket')
-    bracket_div.add(bracket_title)
+#     bracket_div = div(_class='content-container')
+#     bracket_title = h2('Playoff Bracket')
+#     bracket_div.add(bracket_title)
 
-    bracket = main(_id='playoff-bracket')
+#     bracket = main(_id='playoff-bracket')
 
-    rounds = []
+#     rounds = []
 
-    for round in playoff_matchups['Playoff Round'].unique():
-        _ul = ul(_class=f'round round-{round}')
-        _ul.add(li(dominate.util.raw('&nbsp;'), _class='spacer'))
+#     for round in playoff_matchups['Playoff Round'].unique():
+#         _ul = ul(_class=f'round round-{round}')
+#         _ul.add(li(dominate.util.raw('&nbsp;'), _class='spacer'))
 
-        for team1, score1, team2, score2 in playoff_matchups.loc[playoff_matchups['Playoff Round'] == round, ['Home Team','Home Score','Away Team','Away Score']].values:
+#         for team1, score1, team2, score2 in playoff_matchups.loc[playoff_matchups['Playoff Round'] == round, ['Home Team','Home Score','Away Team','Away Score']].values:
 
-            top_team = li(team1, _class='game game-top')
-            if team1 != 'Bye':
-                top_team.add(span(score1))
+#             top_team = li(team1, _class='game game-top')
+#             if team1 != 'Bye':
+#                 top_team.add(span(score1))
 
-            bot_team = li(team2, _class='game game-bottom')
-            if team2 != 'Bye':
-                bot_team.add(span(score2))
+#             bot_team = li(team2, _class='game game-bottom')
+#             if team2 != 'Bye':
+#                 bot_team.add(span(score2))
 
-            _ul.add(top_team)
-            _ul.add(li(dominate.util.raw('&nbsp;'), _class='game game-spacer'))
-            _ul.add(bot_team)
-            _ul.add(li(dominate.util.raw('&nbsp;'), _class='spacer'))
+#             _ul.add(top_team)
+#             _ul.add(li(dominate.util.raw('&nbsp;'), _class='game game-spacer'))
+#             _ul.add(bot_team)
+#             _ul.add(li(dominate.util.raw('&nbsp;'), _class='spacer'))
 
-        rounds.append(_ul)
+#         rounds.append(_ul)
 
-    bracket.add(rounds)
-    bracket_div.add(bracket)
+#     bracket.add(rounds)
+#     bracket_div.add(bracket)
     
-    container.add(summary_div)
-    container.add(scatter_div)
-    container.add(bracket_div)
+#     container.add(summary_div)
+#     container.add(scatter_div)
+#     container.add(bracket_div)
 
-    return container
+#     return container
 
 
 
 def document() -> dominate.document:
     doc = dominate.document(title='Fantasy Football')
-    doc.head.add(link(rel='stylesheet', href=f'{ROOT}style.css'))
-    doc.head.add(link(rel='icon', type='image/png', href=f'{ROOT}Assets/Fantasy-Football-App-LOGO.png'))
+    doc.head.add(link(rel='stylesheet', href=f'{constants.ROOT}style.css'))
+    doc.head.add(link(rel='icon', type='image/png', href=f'{constants.ROOT}Assets/Fantasy-Football-App-LOGO.png'))
 
     return doc
 
@@ -384,62 +385,58 @@ def week_pages():
 
         for week in weeks:
             doc = document()
-            doc.add(header(active_year=year))
+            doc.add(page_header(active_year=year))
 
             doc.add(week_content(year=year, week=week))
 
-            with open(f'fantasy-football-website/seasons/{year}/week-{week}.html','w') as file:
+            with open(f'seasons/{year}/week-{week}.html','w') as file:
                 file.write(doc.render())
 
 # Construct and write the home page
-def home_page():
-    doc = document()
-    doc.add(header(active_year='home'))
+# def home_page():
+#     doc = document()
+#     doc.add(page_header(active_year='home'))
 
-    doc.add(home_content())
+#     doc.add(home_content())
 
-    with open(f'fantasy-football-website/index.html','w') as file:
-        file.write(doc.render())
+#     with open(f'index.html','w') as file:
+#         file.write(doc.render())
 
 # Construct and write the champion page
-def champion_page():
-    doc = document()
-    doc.add(header(active_year='champion'))
+# def champion_page():
+#     doc = document()
+#     doc.add(page_header(active_year='champion'))
 
-    doc.add(champion_content())
+#     doc.add(champion_content())
 
-    with open(f'fantasy-football-website/champion.html','w') as file:
-        file.write(doc.render())
+#     with open(f'champion.html','w') as file:
+#         file.write(doc.render())
 
 # Construct and write the team pages
-def team_pages():
-    conn = sqlite3.connect('fantasy-football-website/database/fantasy-football.db')
+# def team_pages():
+#     for team in constants.TEAMS:
+#         doc = document()
+#         doc.add(page_header(active_year='team'))
 
-    for team in TEAMS:
-        doc = document()
-        doc.add(header(active_year='team'))
+#         doc.add(team_content(team=team))
 
-        doc.add(team_content(team=team, conn=conn))
+#         with open(f'teams/{team}.html','w') as file:
+#             file.write(doc.render())
 
-        with open(f'fantasy-football-website/teams/{team}.html','w') as file:
-            file.write(doc.render())
+# # Construct and write the year pages
+# def year_pages():
+#     for year in constants.YEARS:
+#         doc = document()
+#         doc.add(page_header(active_year=year))
 
-    conn.close()
+#         doc.add(year_content(year))
 
-# Construct and write the year pages
-def year_pages():
-    for year in YEARS:
-        doc = document()
-        doc.add(header(active_year=year))
-
-        doc.add(year_content(year))
-
-        with open(f'fantasy-football-website/seasons/{year}/index.html','w') as file:
-            file.write(doc.render())
+#         with open(f'seasons/{year}/index.html','w') as file:
+#             file.write(doc.render())
 
 # Call the constructing functions
-# home_page()
-# champion_page()
+# home_page.home_page()
+champion_page.champion_page()
 # week_pages()
-# team_pages()
-# year_pages()
+# team_page.team_pages()
+# year_page.year_pages()
