+++ python/document.py
@@ -1,25 +1,14 @@
 import dominate
 from dominate.tags import *
 
-ROOT = '/fantasy-football-website/'
 
-class Document(dominate.document):
-    def __init__(self, title):
-        super().__init__(title)
+from python import constants
+# Only use the below import when debugging this file directly
+# import constants
 
-        self.head.add(link(rel='stylesheet', href=f'{ROOT}style.css'))
-        self.head.add(link(rel='icon', type='image/png', href=f'{ROOT}Assets/Fantasy-Football-App-LOGO.png'))
+def document() -> dominate.document:
+    doc = dominate.document(title='Fantasy Football')
+    doc.head.add(link(rel='stylesheet', href=f'{constants.ROOT}style.css'))
+    doc.head.add(link(rel='icon', type='image/png', href=f'{constants.ROOT}Assets/Fantasy-Football-App-LOGO.png'))
 
-    def print_html(self):
-        print(self.render())
-
-    def write_html(self):
-        pass
-    
-
-
-d = Document('Fantasy Football')
-
-d.add(div(p('test')))
-
-d.print_html()
+    return doc
